Components de cards, logout e login corrigidos
