+++ semana10/spacex/src/pages/TripDetailsPage.js
@@ -3,78 +3,25 @@ import useTrips from "../hooks/useTrips"
 import { useHistory,useParams } from "react-router-dom";
 import { ButtonBack}  from "../components/Buttons/ButtonBack"
 import useProtectedPage from "../hooks/useProtectedPage"
-import Card from "../components/Cards/Cards"
 import { AreaCard } from './style';
 import { BASE_URL, headers } from '../constants/url';
 import axios from "axios"
-import loading from "../Image/loading.svg"
+
 import { ButtonArea } from "./style"
+import CandidateCard from '../components/Cards/candidateCard';
+import ApprovedCard from '../components/Cards/approvedCard';
 
 
 function TripDetailsPage(props) {
+    const params = useParams()
 
     useProtectedPage()
 
+    const Trips = useTrips("/trip",
+    params.id,
+    {})
+
     const history = useHistory()
-    const params = useParams()
-
-        const Trips = useTrips("/trip",
-            params.id,
-            {})
-
-        console.log(Trips)
-
-        const decideCandidate = (id) => {
-            const body = {approve: true}
-
-            axios
-            .put(`${BASE_URL}/trips/${params.id}/candidates/${id}/decide`,body,headers)
-            .then((response) => {
-                alert(response.data.message)
-            })
-            .catch((err) => {
-                console.log(err)
-            })
-        }
-
-        const listCandidates = Trips.candidates && Trips.candidates.map((candidate) => {
-            return(
-
-                    <Card
-                    name = {candidate.name}
-                    description = {candidate.applicationText}
-                    place = {candidate.country}
-                    ocupation = {candidate.profession}
-                    date = {candidate.age}
-                    textButton = "Aprovar"   
-                    actionButton = {() => decideCandidate(candidate.id)}        
-                    />
-
-            )
-        })
-
-        const listApproved = Trips.approved && Trips.approved.map((candidate) => {
-            
-            return(
-                <Card
-                name = {candidate.name}
-                description = {candidate.applicationText}
-                place = {candidate.country}
-                ocupation = {candidate.profession}
-                date = {candidate.age}
-                textButton = "Aprovar"   
-                actionButton = {() => decideCandidate(candidate.id)}        
-                />
-            )
-
-        })
-
-
-        if(!Trips.candidates){
-            return(
-                <img src= {loading}></img> 
-            )
-        }
 
     return (
         <div>
@@ -84,10 +31,14 @@ function TripDetailsPage(props) {
             <p>{Trips.date}</p>
             <p>{Trips.durationInDays}</p>
             <AreaCard>
-                {listCandidates}
+                <CandidateCard
+                    params = {params}
+                />
             </AreaCard> 
                 {Trips.approved && Trips.approved.length > 0 && <h1>Aprovados</h1>}
-                {listApproved}
+                <ApprovedCard
+                    params = {params}
+                />
             <ButtonArea>
                 <ButtonBack
                     h = {history}
